include: arch: arm: Add shared memory for multiple images
Adds a Kconfig option for enabling shared RAM between multiple images
and to the arm linker script. It's calculated in sizes of 1Kb and adjust
the available RAM for the image accordingly. It grows from top down
meaning that the shared RAM section is added to the end of the available
ram.